#940 - Updating project state requires higher privileges
- Move recalculateProjectState from DocumentServiceImpl to ProjectService
- Do not require special permissions for recalculateProjectState
