+++ webanno-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/ProjectService.java
@@ -174,6 +174,14 @@ public interface ProjectService
     void updateProject(Project project);
 
     /**
+     * Update the project state.
+     * 
+     * @param aProject
+     *            The {@link Project} to be updated.
+     */
+    void recalculateProjectState(Project aProject);
+
+    /**
      * A method that check is a project exists with the same name already. getSingleResult() fails
      * if the project is not created, hence existProject returns false.
      *
