+++ webanno-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/ProjectServiceImpl.java
@@ -37,6 +37,7 @@ import java.util.Enumeration;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map.Entry;
+import java.util.Objects;
 import java.util.Properties;
 import java.util.Set;
 import java.util.zip.ZipEntry;
@@ -68,11 +69,14 @@ import de.tudarmstadt.ukp.clarin.webanno.api.ProjectType;
 import de.tudarmstadt.ukp.clarin.webanno.api.SecurityUtil;
 import de.tudarmstadt.ukp.clarin.webanno.api.event.AfterProjectCreatedEvent;
 import de.tudarmstadt.ukp.clarin.webanno.api.event.BeforeProjectRemovedEvent;
+import de.tudarmstadt.ukp.clarin.webanno.api.event.ProjectStateChangedEvent;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
 import de.tudarmstadt.ukp.clarin.webanno.model.PermissionLevel;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.model.ProjectPermission;
+import de.tudarmstadt.ukp.clarin.webanno.model.ProjectState;
+import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.Authority;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
@@ -132,6 +136,71 @@ public class ProjectServiceImpl
     public void updateProject(Project aProject)
     {
         entityManager.merge(aProject);
+    }
+    
+    @Override
+    @Transactional
+    public void recalculateProjectState(Project aProject)
+    {
+        Project project;
+        try {
+            project = getProject(aProject.getId());
+        }
+        catch (NoResultException e) {
+            // This happens when this method is called as part of deleting an entire project.
+            // In such a case, the project may no longer be available, so there is no point in
+            // updating its state. So then we do nothing here.
+            return;
+        }
+        
+        String query = 
+                "SELECT new " + SourceDocumentStateStats.class.getName() + "(" +
+                "COUNT(*), " +
+                "SUM(CASE WHEN state = :an  THEN 1 ELSE 0 END), " +
+                "SUM(CASE WHEN (state = :aip OR state is NULL) THEN 1 ELSE 0 END), " +
+                "SUM(CASE WHEN state = :af  THEN 1 ELSE 0 END), " +
+                "SUM(CASE WHEN state = :cip THEN 1 ELSE 0 END), " +
+                "SUM(CASE WHEN state = :cf  THEN 1 ELSE 0 END)) " +
+                "FROM SourceDocument " + 
+                "WHERE project = :project";
+        
+        SourceDocumentStateStats stats = entityManager.createQuery(
+                        query, SourceDocumentStateStats.class)
+                .setParameter("project", aProject)
+                .setParameter("an", SourceDocumentState.NEW)
+                .setParameter("aip", SourceDocumentState.ANNOTATION_IN_PROGRESS)
+                .setParameter("af", SourceDocumentState.ANNOTATION_FINISHED)
+                .setParameter("cip", SourceDocumentState.CURATION_IN_PROGRESS)
+                .setParameter("cf", SourceDocumentState.CURATION_FINISHED)
+                .getSingleResult();
+        
+        ProjectState oldState = project.getState();
+        
+        if (stats.total == stats.cf) {
+            project.setState(ProjectState.CURATION_FINISHED);
+        }
+        else if (stats.total == stats.af) {
+            project.setState(ProjectState.ANNOTATION_FINISHED);
+        }
+        else if (stats.total == stats.an) {
+            project.setState(ProjectState.NEW);
+        }
+        else if (stats.cip > 0) {
+            project.setState(ProjectState.CURATION_IN_PROGRESS);
+        }
+        else if (stats.aip > 0) {
+            project.setState(ProjectState.ANNOTATION_IN_PROGRESS);
+        }
+        else {
+            throw new IllegalStateException("Unable to determine project state.");
+        }
+        
+        if (!Objects.equals(oldState, project.getState())) {
+            applicationEventPublisher.publishEvent(
+                    new ProjectStateChangedEvent(this, project, oldState));
+        }
+        
+        updateProject(project);
     }
 
     @Override
@@ -898,5 +967,27 @@ public class ProjectServiceImpl
     public List<ProjectType> listProjectTypes()
     {
         return Collections.unmodifiableList(projectTypes);
+    }
+    
+    public static final class SourceDocumentStateStats
+    {
+        public final long total;
+        public final long an;
+        public final long aip;
+        public final long af;
+        public final long cip;
+        public final long cf;
+        
+        public SourceDocumentStateStats(Long aTotal, Long aAn, Long aAip, Long aAf, Long aCip,
+                Long aCf)
+        {
+            super();
+            total = aTotal != null ? aTotal : 0l;
+            an = aAn != null ? aAn : 0l;
+            aip = aAip != null ? aAip : 0l;
+            af = aAf != null ? aAf : 0l;
+            cip = aCip != null ? aCip : 0l;
+            cf = aCf != null ? aCf : 0l;
+        }
     }
 }
